refactored manipulation blueprints for better organization (#2553)
