+++ dimos/control/blueprints/mobile.py
@@ -26,10 +26,10 @@ from __future__ import annotations
 
 import os
 
-from dimos.control.blueprints._hardware import mock_arm
 from dimos.control.components import (
     HardwareComponent,
     HardwareType,
+    make_joints,
     make_twist_base_joints,
 )
 from dimos.control.coordinator import ControlCoordinator, TaskConfig
@@ -195,7 +195,12 @@ coordinator_flowbase_nav = (
 
 
 # Mock arm (7-DOF) + mock holonomic base (3-DOF)
-_mock_arm_hw = mock_arm("arm", 7)
+_mock_arm_hw = HardwareComponent(
+    hardware_id="arm",
+    hardware_type=HardwareType.MANIPULATOR,
+    joints=make_joints("arm", 7),
+    adapter_type="mock",
+)
 
 coordinator_mobile_manip_mock = ControlCoordinator.blueprint(
     hardware=[_mock_arm_hw, _mock_twist_base()],
