+++ dimos/robot/all_blueprints.py
@@ -18,31 +18,31 @@
 all_blueprints = {
     "alfred-nav": "dimos.robot.diy.alfred.blueprints.alfred_nav:alfred_nav",
     "coordinator-basic": "dimos.control.blueprints.basic:coordinator_basic",
-    "coordinator-cartesian-ik-mock": "dimos.control.blueprints.teleop:coordinator_cartesian_ik_mock",
-    "coordinator-cartesian-ik-piper": "dimos.control.blueprints.teleop:coordinator_cartesian_ik_piper",
-    "coordinator-combined-xarm6": "dimos.control.blueprints.teleop:coordinator_combined_xarm6",
-    "coordinator-dual-mock": "dimos.control.blueprints.dual:coordinator_dual_mock",
-    "coordinator-dual-xarm": "dimos.control.blueprints.dual:coordinator_dual_xarm",
+    "coordinator-cartesian-ik-mock": "dimos.robot.manipulators.piper.blueprints.teleop:coordinator_cartesian_ik_mock",
+    "coordinator-cartesian-ik-piper": "dimos.robot.manipulators.piper.blueprints.teleop:coordinator_cartesian_ik_piper",
+    "coordinator-combined-xarm6": "dimos.robot.manipulators.xarm.blueprints.teleop:coordinator_combined_xarm6",
+    "coordinator-dual-mock": "dimos.robot.manipulators.common.mock:coordinator_dual_mock",
+    "coordinator-dual-xarm": "dimos.robot.manipulators.xarm.blueprints.basic:coordinator_dual_xarm",
     "coordinator-flowbase": "dimos.control.blueprints.mobile:coordinator_flowbase",
     "coordinator-flowbase-keyboard-teleop": "dimos.control.blueprints.mobile:coordinator_flowbase_keyboard_teleop",
     "coordinator-flowbase-nav": "dimos.control.blueprints.mobile:coordinator_flowbase_nav",
     "coordinator-mobile-manip-mock": "dimos.control.blueprints.mobile:coordinator_mobile_manip_mock",
-    "coordinator-mock": "dimos.control.blueprints.basic:coordinator_mock",
+    "coordinator-mock": "dimos.robot.manipulators.common.mock:coordinator_mock",
     "coordinator-mock-twist-base": "dimos.control.blueprints.mobile:coordinator_mock_twist_base",
-    "coordinator-openarm-bimanual": "dimos.robot.manipulators.openarm.blueprints:coordinator_openarm_bimanual",
-    "coordinator-openarm-left": "dimos.robot.manipulators.openarm.blueprints:coordinator_openarm_left",
-    "coordinator-openarm-mock": "dimos.robot.manipulators.openarm.blueprints:coordinator_openarm_mock",
-    "coordinator-openarm-right": "dimos.robot.manipulators.openarm.blueprints:coordinator_openarm_right",
-    "coordinator-piper": "dimos.control.blueprints.basic:coordinator_piper",
-    "coordinator-piper-xarm": "dimos.control.blueprints.dual:coordinator_piper_xarm",
-    "coordinator-servo-xarm6": "dimos.control.blueprints.teleop:coordinator_servo_xarm6",
-    "coordinator-teleop-dual": "dimos.control.blueprints.teleop:coordinator_teleop_dual",
-    "coordinator-teleop-piper": "dimos.control.blueprints.teleop:coordinator_teleop_piper",
-    "coordinator-teleop-xarm6": "dimos.control.blueprints.teleop:coordinator_teleop_xarm6",
-    "coordinator-teleop-xarm7": "dimos.control.blueprints.teleop:coordinator_teleop_xarm7",
-    "coordinator-velocity-xarm6": "dimos.control.blueprints.teleop:coordinator_velocity_xarm6",
-    "coordinator-xarm6": "dimos.control.blueprints.basic:coordinator_xarm6",
-    "coordinator-xarm7": "dimos.control.blueprints.basic:coordinator_xarm7",
+    "coordinator-openarm-bimanual": "dimos.robot.manipulators.openarm.blueprints.basic:coordinator_openarm_bimanual",
+    "coordinator-openarm-left": "dimos.robot.manipulators.openarm.blueprints.basic:coordinator_openarm_left",
+    "coordinator-openarm-mock": "dimos.robot.manipulators.openarm.blueprints.basic:coordinator_openarm_mock",
+    "coordinator-openarm-right": "dimos.robot.manipulators.openarm.blueprints.basic:coordinator_openarm_right",
+    "coordinator-piper": "dimos.robot.manipulators.piper.blueprints.basic:coordinator_piper",
+    "coordinator-piper-xarm": "dimos.robot.manipulators.common.mixed:coordinator_piper_xarm",
+    "coordinator-servo-xarm6": "dimos.robot.manipulators.xarm.blueprints.teleop:coordinator_servo_xarm6",
+    "coordinator-teleop-dual": "dimos.robot.manipulators.common.mixed:coordinator_teleop_dual",
+    "coordinator-teleop-piper": "dimos.robot.manipulators.piper.blueprints.teleop:coordinator_teleop_piper",
+    "coordinator-teleop-xarm6": "dimos.robot.manipulators.xarm.blueprints.teleop:coordinator_teleop_xarm6",
+    "coordinator-teleop-xarm7": "dimos.robot.manipulators.xarm.blueprints.teleop:coordinator_teleop_xarm7",
+    "coordinator-velocity-xarm6": "dimos.robot.manipulators.xarm.blueprints.teleop:coordinator_velocity_xarm6",
+    "coordinator-xarm6": "dimos.robot.manipulators.xarm.blueprints.basic:coordinator_xarm6",
+    "coordinator-xarm7": "dimos.robot.manipulators.xarm.blueprints.basic:coordinator_xarm7",
     "demo-agent": "dimos.agents.demo_agent:demo_agent",
     "demo-agent-camera": "dimos.agents.demo_agent:demo_agent_camera",
     "demo-camera": "dimos.hardware.sensors.camera.module:demo_camera",
@@ -58,13 +58,13 @@ all_blueprints = {
     "desk-marker-tf": "dimos.perception.fiducial.blueprints.desk_marker_tf:desk_marker_tf",
     "drone-agentic": "dimos.robot.drone.blueprints.agentic.drone_agentic:drone_agentic",
     "drone-basic": "dimos.robot.drone.blueprints.basic.drone_basic:drone_basic",
-    "dual-xarm6-planner": "dimos.manipulation.blueprints:dual_xarm6_planner",
-    "keyboard-teleop-a750": "dimos.robot.manipulators.a750.blueprints:keyboard_teleop_a750",
-    "keyboard-teleop-openarm": "dimos.robot.manipulators.openarm.blueprints:keyboard_teleop_openarm",
-    "keyboard-teleop-openarm-mock": "dimos.robot.manipulators.openarm.blueprints:keyboard_teleop_openarm_mock",
-    "keyboard-teleop-piper": "dimos.robot.manipulators.piper.blueprints:keyboard_teleop_piper",
-    "keyboard-teleop-xarm6": "dimos.robot.manipulators.xarm.blueprints:keyboard_teleop_xarm6",
-    "keyboard-teleop-xarm7": "dimos.robot.manipulators.xarm.blueprints:keyboard_teleop_xarm7",
+    "dual-xarm6-planner": "dimos.robot.manipulators.xarm.blueprints.basic:dual_xarm6_planner",
+    "keyboard-teleop-a750": "dimos.robot.manipulators.a750.blueprints.teleop:keyboard_teleop_a750",
+    "keyboard-teleop-openarm": "dimos.robot.manipulators.openarm.blueprints.teleop:keyboard_teleop_openarm",
+    "keyboard-teleop-openarm-mock": "dimos.robot.manipulators.openarm.blueprints.teleop:keyboard_teleop_openarm_mock",
+    "keyboard-teleop-piper": "dimos.robot.manipulators.piper.blueprints.teleop:keyboard_teleop_piper",
+    "keyboard-teleop-xarm6": "dimos.robot.manipulators.xarm.blueprints.teleop:keyboard_teleop_xarm6",
+    "keyboard-teleop-xarm7": "dimos.robot.manipulators.xarm.blueprints.teleop:keyboard_teleop_xarm7",
     "mid360": "dimos.hardware.sensors.lidar.livox.livox_blueprints:mid360",
     "mid360-fastlio": "dimos.hardware.sensors.lidar.fastlio2.fastlio_blueprints:mid360_fastlio",
     "mid360-fastlio-ray-trace": "dimos.hardware.sensors.lidar.fastlio2.fastlio_blueprints:mid360_fastlio_ray_trace",
@@ -72,8 +72,8 @@ all_blueprints = {
     "mid360-fastlio-voxels-native": "dimos.hardware.sensors.lidar.fastlio2.fastlio_blueprints:mid360_fastlio_voxels_native",
     "mid360-pointlio": "dimos.hardware.sensors.lidar.pointlio.pointlio_blueprints:mid360_pointlio",
     "mid360-pointlio-voxels": "dimos.hardware.sensors.lidar.pointlio.pointlio_blueprints:mid360_pointlio_voxels",
-    "openarm-mock-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints:openarm_mock_planner_coordinator",
-    "openarm-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints:openarm_planner_coordinator",
+    "openarm-mock-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints.planner:openarm_mock_planner_coordinator",
+    "openarm-planner-coordinator": "dimos.robot.manipulators.openarm.blueprints.planner:openarm_planner_coordinator",
     "path-planner-eval": "dimos.navigation.nav_3d.evaluator.blueprints:path_planner_eval",
     "teleop-hosted-go2": "dimos.teleop.quest_hosted.blueprints:teleop_hosted_go2",
     "teleop-hosted-xarm7": "dimos.teleop.quest_hosted.blueprints:teleop_hosted_xarm7",
@@ -123,13 +123,13 @@ all_blueprints = {
     "unitree-go2-webrtc-keyboard-teleop": "dimos.robot.unitree.go2.blueprints.basic.unitree_go2_webrtc_keyboard_teleop:unitree_go2_webrtc_keyboard_teleop",
     "unitree-go2-webrtc-rage-keyboard-teleop": "dimos.robot.unitree.go2.blueprints.basic.unitree_go2_webrtc_rage_keyboard_teleop:unitree_go2_webrtc_rage_keyboard_teleop",
     "unity-sim": "dimos.simulation.unity.blueprint:unity_sim",
-    "xarm-perception": "dimos.manipulation.blueprints:xarm_perception",
-    "xarm-perception-agent": "dimos.manipulation.blueprints:xarm_perception_agent",
-    "xarm-perception-sim": "dimos.manipulation.blueprints:xarm_perception_sim",
-    "xarm-perception-sim-agent": "dimos.manipulation.blueprints:xarm_perception_sim_agent",
-    "xarm6-planner-only": "dimos.manipulation.blueprints:xarm6_planner_only",
-    "xarm7-planner-coordinator": "dimos.manipulation.blueprints:xarm7_planner_coordinator",
-    "xarm7-planner-coordinator-agent": "dimos.manipulation.blueprints:xarm7_planner_coordinator_agent",
+    "xarm-perception": "dimos.robot.manipulators.xarm.blueprints.perception:xarm_perception",
+    "xarm-perception-agent": "dimos.robot.manipulators.xarm.blueprints.agentic:xarm_perception_agent",
+    "xarm-perception-sim": "dimos.robot.manipulators.xarm.blueprints.simulation:xarm_perception_sim",
+    "xarm-perception-sim-agent": "dimos.robot.manipulators.xarm.blueprints.agentic:xarm_perception_sim_agent",
+    "xarm6-planner-only": "dimos.robot.manipulators.xarm.blueprints.basic:xarm6_planner_only",
+    "xarm7-planner-coordinator": "dimos.robot.manipulators.xarm.blueprints.basic:xarm7_planner_coordinator",
+    "xarm7-planner-coordinator-agent": "dimos.robot.manipulators.xarm.blueprints.agentic:xarm7_planner_coordinator_agent",
 }
