+++ dimos/robot/manipulators/a750/blueprints/teleop.py
@@ -1,0 +1,51 @@
+# Copyright 2026 Dimensional Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""A-750 teleop blueprints."""
+
+from __future__ import annotations
+
+from dimos.control.coordinator import ControlCoordinator
+from dimos.core.coordination.blueprints import autoconnect
+from dimos.manipulation.manipulation_module import ManipulationModule
+from dimos.robot.manipulators.a750.config import (
+    A750_FK_MODEL,
+    A750_HOME_JOINTS,
+    a750_hardware,
+    make_a750_model_config,
+)
+from dimos.robot.manipulators.common.blueprints import cartesian_ik_task
+from dimos.teleop.keyboard.keyboard_teleop_module import KeyboardTeleopModule
+
+_a750_hw = a750_hardware("arm", mock_without_address=True)
+
+keyboard_teleop_a750 = autoconnect(
+    KeyboardTeleopModule.blueprint(
+        model_path=A750_FK_MODEL,
+        ee_joint_id=6,
+        home_joints=A750_HOME_JOINTS,
+        joint_names=_a750_hw.joints,
+    ),
+    ControlCoordinator.blueprint(
+        tick_rate=100.0,
+        publish_joint_state=True,
+        joint_state_frame_id="coordinator",
+        hardware=[_a750_hw],
+        tasks=[cartesian_ik_task(_a750_hw, model_path=A750_FK_MODEL, ee_joint_id=6)],
+    ),
+    ManipulationModule.blueprint(
+        robots=[make_a750_model_config()],
+        visualization={"backend": "meshcat"},
+    ),
+)
