+++ dimos/robot/manipulators/openarm/blueprints/teleop.py
@@ -1,0 +1,71 @@
+# Copyright 2026 Dimensional Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""OpenArm keyboard teleop blueprints."""
+
+from __future__ import annotations
+
+from dimos.control.coordinator import ControlCoordinator
+from dimos.core.coordination.blueprints import autoconnect
+from dimos.manipulation.manipulation_module import ManipulationModule
+from dimos.robot.manipulators.common.blueprints import cartesian_ik_task
+from dimos.robot.manipulators.openarm.config import (
+    LEFT_CAN,
+    OPENARM_V10_FK_MODEL,
+    openarm_single_hardware,
+    openarm_single_model_config,
+)
+from dimos.teleop.keyboard.keyboard_teleop_module import KeyboardTeleopModule
+
+_teleop_hw = openarm_single_hardware()
+
+keyboard_teleop_openarm_mock = autoconnect(
+    KeyboardTeleopModule.blueprint(
+        model_path=OPENARM_V10_FK_MODEL,
+        ee_joint_id=7,
+        joint_names=_teleop_hw.joints,
+    ),
+    ControlCoordinator.blueprint(
+        hardware=[_teleop_hw],
+        tasks=[cartesian_ik_task(_teleop_hw, model_path=OPENARM_V10_FK_MODEL, ee_joint_id=7)],
+    ),
+    ManipulationModule.blueprint(
+        robots=[openarm_single_model_config()],
+        visualization={"backend": "meshcat"},
+    ),
+)
+
+_teleop_real_hw = openarm_single_hardware(adapter_type="openarm", address=LEFT_CAN)
+
+keyboard_teleop_openarm = autoconnect(
+    KeyboardTeleopModule.blueprint(
+        model_path=OPENARM_V10_FK_MODEL,
+        ee_joint_id=7,
+        joint_names=_teleop_real_hw.joints,
+    ),
+    ControlCoordinator.blueprint(
+        hardware=[_teleop_real_hw],
+        tasks=[
+            cartesian_ik_task(
+                _teleop_real_hw,
+                model_path=OPENARM_V10_FK_MODEL,
+                ee_joint_id=7,
+            )
+        ],
+    ),
+    ManipulationModule.blueprint(
+        robots=[openarm_single_model_config()],
+        visualization={"backend": "meshcat"},
+    ),
+)
