+++ dimos/teleop/quest/blueprints.py
@@ -20,17 +20,17 @@ hardware. The underlying coordinator blueprints branch on `global_config.simulat
 """
 
 from dimos.constants import DEFAULT_CAPACITY_COLOR_IMAGE
-from dimos.control.blueprints.teleop import (
-    coordinator_teleop_dual,
-    coordinator_teleop_piper,
-    coordinator_teleop_xarm6,
-    coordinator_teleop_xarm7,
-)
 from dimos.core.coordination.blueprints import autoconnect
 from dimos.core.transport import LCMTransport, pSHMTransport
 from dimos.msgs.geometry_msgs.PoseStamped import PoseStamped
 from dimos.msgs.geometry_msgs.Twist import Twist
 from dimos.msgs.sensor_msgs.Image import Image
+from dimos.robot.manipulators.common.mixed import coordinator_teleop_dual
+from dimos.robot.manipulators.piper.blueprints.teleop import coordinator_teleop_piper
+from dimos.robot.manipulators.xarm.blueprints.teleop import (
+    coordinator_teleop_xarm6,
+    coordinator_teleop_xarm7,
+)
 from dimos.robot.unitree.go2.connection import GO2Connection
 from dimos.teleop.quest.quest_extensions import (
     ArmTeleopModule,
