+++ docs/capabilities/manipulation/a750.md
@@ -49,8 +49,10 @@ DEVICE_PATH=/dev/ttyACM0 dimos run keyboard-teleop-a750
 
 ## Robot Model
 
-The robot model, hardware, and task config used by the runnable stack are
-defined in [`dimos/robot/manipulators/a750/blueprints.py`](/dimos/robot/manipulators/a750/blueprints.py).
+The robot model and hardware config are defined in
+[`dimos/robot/manipulators/a750/config.py`](/dimos/robot/manipulators/a750/config.py).
+The runnable keyboard teleop stack is composed in
+[`dimos/robot/manipulators/a750/blueprints/teleop.py`](/dimos/robot/manipulators/a750/blueprints/teleop.py).
 
 | Field | Value |
 |-------|-------|
@@ -108,7 +110,8 @@ If `a750_control` is not installed, `connect()` records an error and returns `Fa
 
 ## Control Path
 
-The `keyboard-teleop-a750` blueprint is defined in [`dimos/robot/manipulators/a750/blueprints.py`](/dimos/robot/manipulators/a750/blueprints.py#L22).
+The `keyboard-teleop-a750` blueprint is defined in
+[`dimos/robot/manipulators/a750/blueprints/teleop.py`](/dimos/robot/manipulators/a750/blueprints/teleop.py).
 
 ```text
 KeyboardTeleopModule
